make repair into double and summary scripts

+++ src/dynamic_programming/stress_damage.cpp
@@ -49,7 +49,7 @@ const double hslope   = 20.0;     // slope parameter controlling increase in dam
 
 // damage units removed per time step
 // this can be (re)set through the command line
-int repair      = 1;      
+double repair      = 1;      
 
 
 const double K        = 0.001;   // parameter K controlling increase in mortality with damage level
@@ -158,7 +158,7 @@ void Damage()
   {
     for (h=0;h<=maxH;h++)
     {
-      dnew[d][h] = max(0.0,min(double(maxD),double(d) + hslope*(hmin-(double(h)/double(maxH)))*(hmin-(double(h)/double(maxH))) - double(repair)));
+      dnew[d][h] = max(0.0,min(double(maxD),double(d) + hslope*(hmin-(double(h)/double(maxH)))*(hmin-(double(h)/double(maxH))) - repair));
     }
   }
 }
@@ -561,7 +561,7 @@ int main(int argc, char **argv)
 {
     double autocorr = atof(argv[1]);
     double risk = atof(argv[2]);
-    repair = atoi(argv[3]);
+    repair = atof(argv[3]);
 
         pLeave = (1.0 - autocorr)/(1.0+(risk/(1.0-risk)));
         pArrive = 1.0 - pLeave - autocorr;
